+++ gen/com/example/saveme_android/BuildConfig.java
@@ -1,0 +1,6 @@
+/** Automatically generated file. DO NOT MODIFY */
+package com.example.saveme_android;
+
+public final class BuildConfig {
+    public final static boolean DEBUG = true;
+}
