+++ res/values/strings.xml
@@ -1,4 +1,5 @@
 <resources>
-    <string name="app_name">GoogleDrive</string>
-    <string name="action_settings">Settings</string>
+
+    <string name="app_name">Saveme</string>
+
 </resources>
